Add BST :evergreen_tree:

+++ Data-Structures/Array-Structures/circular_deque.cpp
@@ -15,15 +15,15 @@
 template <typename T>
 class CircularDeque final {
 public:
-    static constexpr size_t DEFAULT_CAPACITY = 10;
-    static constexpr size_t DEFAULT_CAPACITY_MULTIPLIER = 2;
-    explicit CircularDeque(const T& initialValue, const size_t& capacity = DEFAULT_CAPACITY) : _head(0), _tail(0), _size(1), _capacity(capacity), _deque(std::make_unique<T[]>(capacity)) {
+    static constexpr unsigned DEFAULT_CAPACITY = 10;
+    static constexpr unsigned DEFAULT_CAPACITY_MULTIPLIER = 2;
+    explicit CircularDeque(const T& initialValue, const unsigned& capacity = DEFAULT_CAPACITY) : _head(0), _tail(0), _size(1), _capacity(capacity), _deque(std::make_unique<T[]>(capacity)) {
         _deque[_head] = initialValue;
     }
-    size_t getSize() const {
+    unsigned getSize() const {
         return _size;
     }
-    size_t getCapacity() const {
+    unsigned getCapacity() const {
         return _capacity;
     }
     bool empty() const {
@@ -72,21 +72,21 @@ public:
     void print(std::ostream& out = std::cout) const {
         if (empty())
             throw circularDequeEmpty;
-        for (size_t i = _head; i != _tail; (i == _capacity - 1) ? (i = 0) : (++i)) {
+        for (unsigned i = _head; i != _tail; (i == _capacity - 1) ? (i = 0) : (++i)) {
             out << _deque[i] << ' ';
         }
         out << _deque[_tail] << std::endl;
     }
 private:
-    size_t _head;
-    size_t _tail;
-    size_t _size;
-    size_t _capacity;
+    unsigned _head;
+    unsigned _tail;
+    unsigned _size;
+    unsigned _capacity;
     std::unique_ptr<T[]> _deque;
     void _increaseCapacity() {
         _capacity *= DEFAULT_CAPACITY_MULTIPLIER;
         std::unique_ptr<T[]> _newDeque = std::make_unique<T[]>(_capacity);
-        for (size_t i = 0, j = _head; i < _size; ++i, (j == _size - 1) ? (j = 0) : (++j)) {
+        for (unsigned i = 0, j = _head; i < _size; ++i, (j == _size - 1) ? (j = 0) : (++j)) {
             _newDeque[i] = _deque[j];
         }
         _head = 0;
